pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sizzle"
version = "0.1.3"
description = "A pluggable Django app for daily check-ins, time tracking, and team productivity monitoring"
readme = "README.md"
license = "MIT"
authors = [
    {name = "OWASP BLT Team", email = "blt@owasp.org"}
]
maintainers = [
    {name = "OWASP BLT Team", email = "blt@owasp.org"}
]
keywords = ["django", "time-tracking", "daily-checkins", "productivity", "team-management"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1", 
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Framework :: Django :: 5.1",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.11"

# Core dependencies (commented out for TestPyPI testing)
# dependencies = [
#     "Django>=4.0,<6.0"
# ]

# Uncomment for production release:
dependencies = []

# Optional dependencies for different features
[project.optional-dependencies]
# Slack integration support
slack = [
    "slack-bolt>=1.20.0",
    "slack-sdk>=3.30.0",
]

# Email functionality
email = [
    "sendgrid>=6.11.0",
]

# Full feature set (includes all optional dependencies)
full = [
    "slack-bolt>=1.20.0",
    "slack-sdk>=3.30.0",
    "sendgrid>=6.11.0",
]

# Development dependencies
dev = [
    "black>=24.0.0",
    "isort>=5.13.0",
    "ruff>=0.14.0",
    "pre-commit>=3.8.0",
    "django-debug-toolbar>=4.4.0",
]

# Testing dependencies
test = [
    "pytest>=7.0.0",
    "pytest-django>=4.5.0",
    "pytest-cov>=4.0.0",
    "factory-boy>=3.3.0",
    "freezegun>=1.2.0",
]

[project.urls]
Homepage = "https://github.com/OWASP-BLT/BLT"
Documentation = "https://github.com/OWASP-BLT/BLT/tree/main/docs"
Repository = "https://github.com/OWASP-BLT/BLT"
"Bug Tracker" = "https://github.com/OWASP-BLT/BLT/issues"
Changelog = "https://github.com/OWASP-BLT/BLT/blob/main/CHANGELOG.md"

[project.entry-points."django.apps"]
sizzle = "sizzle.apps:SizzleConfig"

[tool.setuptools]
packages = ["sizzle"]
include-package-data = true

[tool.setuptools.package-data]
sizzle = [
    "templates/**/*",
    "static/**/*", 
    "migrations/*.py",
    "management/**/*",
    "locale/**/*"
]

# Code formatting and linting configuration
[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 120
known_first_party = ["sizzle"]
known_django = ["django"]
sections = ["FUTURE", "STDLIB", "DJANGO", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.ruff]
line-length = 120
target-version = "py311"
extend-exclude = ["migrations"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings  
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "DJ", # flake8-django
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "DJ01",  # django model without __str__ method
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"settings.py" = ["F405", "F403"]
"*/migrations/*" = ["E501", "F401", "F841"]

# Django-specific ruff rules
[tool.ruff.lint.flake8-django]
django-settings-module = "tests.settings"

# Coverage configuration for testing
[tool.coverage.run]
source = ["sizzle"]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/venv/*",
    "setup.py",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]

# Pytest configuration
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--verbose --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
